rename process to exec
