+++ src/exec/exec_pipe.c
@@ -1,32 +1,30 @@
 /* ************************************************************************** */
 /*                                                                            */
 /*                                                        :::      ::::::::   */
-/*   execute_outfile_append_process.c                   :+:      :+:    :+:   */
+/*   exec_pipe.c                                        :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
 /*   By: reasuke <reasuke@student.42tokyo.jp>       +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
-/*   Created: 2024/02/25 20:14:25 by reasuke           #+#    #+#             */
-/*   Updated: 2024/03/23 19:54:05 by reasuke          ###   ########.fr       */
+/*   Created: 2024/03/16 16:33:17 by reasuke           #+#    #+#             */
+/*   Updated: 2024/05/27 22:48:11 by reasuke          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
-#include "process.h"
+#include "exec.h"
 #include "wrapper.h"
 
-pid_t	execute_outfile_append_process(const char *outfile_path,
-			const char *cmd, int fds[2], char **envp)
+pid_t	execute_pipe_process(const char *cmd, int prev_fds[2], int next_fds[2],
+			char **envp)
 {
 	pid_t	pid;
-	int		out_fd;
 
 	pid = xfork();
 	if (pid != CHILD)
 		return (pid);
-	out_fd = xopen_with_permission(outfile_path, O_WRONLY | O_CREAT | O_APPEND,
-			0664);
-	xclose(fds[1]);
-	xdup2(fds[0], STDIN_FILENO);
-	xdup2(out_fd, STDOUT_FILENO);
+	xclose(prev_fds[1]);
+	xdup2(prev_fds[0], STDIN_FILENO);
+	xclose(next_fds[0]);
+	xdup2(next_fds[1], STDOUT_FILENO);
 	execute_command(cmd, envp);
 	return (0);
 }
