+++ src/exec/exec_processes.c
@@ -1,17 +1,18 @@
 /* ************************************************************************** */
 /*                                                                            */
 /*                                                        :::      ::::::::   */
-/*   execute_processes.c                                :+:      :+:    :+:   */
+/*   exec_processes.c                                   :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
 /*   By: reasuke <reasuke@student.42tokyo.jp>       +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/03/17 12:38:05 by reasuke           #+#    #+#             */
-/*   Updated: 2024/03/23 19:46:27 by reasuke          ###   ########.fr       */
+/*   Updated: 2024/05/27 22:59:26 by reasuke          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
-#include "process.h"
 #include "libft.h"
+
+#include "exec.h"
 #include "wrapper.h"
 
 static void	_close_pipe_end(int fds[2])
